refactor: simplify MessageWrapper
